+++ runway/__init__.py
@@ -1,2 +1,2 @@
 """Set package version."""
-__version__ = '0.44.3'
+__version__ = '0.45.0'